+++ scripts/ci/sync_all_releases.py
@@ -86,3 +86,4 @@ def sync_releases():
 
 if __name__ == "__main__":
     sync_releases()
+